DecisionTreeAndRandomForest

Documentation for DecisionTreeAndRandomForest.

DecisionTreeAndRandomForest.ClassificationTree

DecisionTreeAndRandomForest.Leaf

DecisionTreeAndRandomForest.Node

DecisionTreeAndRandomForest.best_split

DecisionTreeAndRandomForest.build_tree

DecisionTreeAndRandomForest.entropy

DecisionTreeAndRandomForest.find_best_split

DecisionTreeAndRandomForest.fit

DecisionTreeAndRandomForest.gini_impurity

DecisionTreeAndRandomForest.information_gain

DecisionTreeAndRandomForest.predict

DecisionTreeAndRandomForest.print_tree

DecisionTreeAndRandomForest.split_dataset

DecisionTreeAndRandomForest.split_node

DecisionTreeAndRandomForest.weighted_gini

DecisionTreeAndRandomForest.ClassificationTree — Type

Represents a ClassificationTree.

max_depth controls the maximum depth of the tree. If -1, the depth is not limited.

min_samples_split controls when a node in the decision tree should be split.

root contains the root Node of the ClassificationTree.

data contains the datapoints of the ClassificationTree.

labels contains the respective labels of the datapoints.

DecisionTreeAndRandomForest.Leaf — Type

Represents a Leaf in the ClassificationTree structure.

values stores the labels of the data points.

DecisionTreeAndRandomForest.Node — Type

Represents a Node in the ClassificationTree structure.

left points to the left child.

right points to the right child.

feature_index stores the index of the selected feature.

split_value stores the value on that the data is split.

data contains the datapoints of the Node.

labels contains the respective labels of the datapoints.

DecisionTreeAndRandomForest.best_split — Method

best_split(X::AbstractMatrix{T}, y::Vector{T}) where T

Find the best split for the dataset X and labels y based on Information Gain. Returns the best feature and threshold for the split.

Arguments

X: A matrix of features.

y: A vector of labels.

Returns

best_feature: The index of the best feature to split on.

best_threshold: The threshold value for the best split.

DecisionTreeAndRandomForest.build_tree — Method

build_tree(data, labels, max_depth, min_samples_split, depth)

Build the tree structure of the ClassificationTree

If depth is unspecified, it is set to 0

DecisionTreeAndRandomForest.entropy — Method

entropy(y::Vector{T}) where T

Calculate the entropy of a vector of labels y.

Arguments

y: A vector of labels.

Returns

The entropy of the vector.

DecisionTreeAndRandomForest.find_best_split — Method

Finds the best split point for a decision tree node. For now it uses the Gini impurity as splitting criterion, but should later be extended to support other criteria.

Args: data: A matrix of features, where each row is a data point and each column is a feature. labels: A vector of labels corresponding to the data points.

Returns: A tuple containing the index of the best feature and the best split value.

DecisionTreeAndRandomForest.fit — Method

fit(tree::ClassificationTree)

Compute the tree structure.

DecisionTreeAndRandomForest.gini_impurity — Method

Calculates the Gini impurity of a set of labels.

Args: labels: A vector of labels.

Returns: The Gini impurity of the labels.

DecisionTreeAndRandomForest.information_gain — Method

information_gain(y::Vector{T}, y_left::Vector{T}, y_right::Vector{T}) where T

Calculate the Information Gain of a split.

Arguments

y: The original labels vector.

y_left: The labels vector for the left split.

y_right: The labels vector for the right split.

Returns

The Information Gain of the split.

DecisionTreeAndRandomForest.predict — Method

predict(tree::ClassificationTree, data::Matrix{T})

Returns the prediction of the ClassificationTree for a list of datapoints.

data contains the datapoints to predict.

DecisionTreeAndRandomForest.print_tree — Method

print_tree(tree:ClassificationTree)

Prints the tree structure. Mainly used for debugging purposes.

DecisionTreeAndRandomForest.split_dataset — Method

split_dataset(X::AbstractMatrix{T}, y::Vector{T}, feature::Int, threshold::Real) where T

Split the dataset X and labels y based on a feature and a threshold. Returns the left and right splits for both X and y.

Arguments

X: A matrix of features.

y: A vector of labels.

feature: The index of the feature to split on.

threshold: The threshold value to split the feature.

Returns

X_left, y_left: The left split of the dataset and labels.

X_right, y_right: The right split of the dataset and labels.

DecisionTreeAndRandomForest.split_node — Method

Splits the labels into two nodes based on the provided feature and value.

Args: data: A matrix of features, where each row is a data point and each column is a feature. labels: A vector of labels corresponding to the data points. index: The index of the feature to split on. value: The value to split the feature on.

Returns: A tuple containing the left and right sets of labels.

DecisionTreeAndRandomForest.weighted_gini — Method

Calculates the weighted Gini impurity of a split.

Args: leftdataset: A vector of labels for the left subset of the data. rightdataset: A vector of labels for the right subset of the data.

Returns: The weighted Gini impurity of the split.